pyproject.toml
[tool.poetry]
name = "Feecc-Workbench-Daemon"
version = "2.0"
description = "Workbench software for the Feecc QA system"
authors = ["arseniiarsenii <arseniivelichko2@gmail.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
Pillow = "^9.2.0"
PyYAML = "^6.0"
brother-ql = "^0.9.4"
qrcode = "^7.3.1"
python-barcode = "^0.14.0"
fastapi = "^0.82.0"
uvicorn = "^0.18.3"
dnspython = "^2.2.1"
loguru = "^0.6.0"
motor = "^3.0.0"
httpx = "^0.23.0"
environ-config = "^22.1.0"
yarl = "^1.8.1"
sse-starlette = "^1.1.6"
python-dotenv = "^0.21.0"
robonomics-interface = "^1.2.2"
ecs-logging = "^2.0.0"
aioprometheus = "^22.5.0"
pycups = "^2.0.1"

[tool.poetry.dev-dependencies]
mypy = "^0.971"
flake8 = "^5.0.4"
types-PyYAML = "^6.0.11"
types-requests = "^2.28.9"
pytest = "^7.1.3"
vulture = "^2.5"
pre-commit = "^2.20.0"
pytest-cov = "^3.0.0"
ssort = "^0.11.6"
black = {extras = ["d"], version = "^22.8.0"}
flake8-new-union-types = "^0.4.1"
flake8-pep585 = "^0.1.5.1"
pyupgrade = "^2.37.3"
isort = "^5.10.1"
poetryup = "^0.10.0"

[tool.poetry.group.dev.dependencies]
flake8-json = "^21.7.0"
flake8-adjustable-complexity = "^0.0.6"
flake8-annotations-coverage = "^0.0.6"
flake8-builtins = "^1.5.3"
flake8-cognitive-complexity = "^0.1.0"
flake8-comprehensions = "^3.10.0"
flake8-expression-complexity = "^0.0.11"
flake8-functions = "^0.0.7"
flake8-simplify = "^0.19.3"
flake8-scream = "^0.1.0"
pep8-naming = "^0.13.2"
flake8-bugbear = "^22.8.23"
flake8-print = "^5.0.0"
flake8-use-pathlib = "^0.3.0"
pylint = "^2.15.2"
flake8-pylint = "^0.1.3"

[tool.pytest.ini_options]
markers = ["short_url"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = "test*"
strict = true
ignore_missing_imports  = true
allow_untyped_decorators = true
implicit_reexport = true

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
